Updated to version 1.0.6
Reduced the number of “design” in config.json to 2.
Changed the courses.json so that the array also uses “number” and
“qualityHours”.
Updated favicon.ico to UWI crest.
Updated main.html to read “The Office of the Board”, copyright date
will now be updated automatically.
Updated version4.html and addcoursesmodal.html so that Grade comes
before Level.
Except when faculty of Engineering is selected, selecting a course will
also select the corresponding level and credit hours.